+++ src/residue.js
@@ -38,9 +38,6 @@ const Params = {
     // whether connection is being made (enabled at connection_socket, disabled at logging_socket)
     connecting: false,
 
-    // lock for mutex
-    lock: false,
-
     // list of tokens currently available
     tokens: [],
 
@@ -51,7 +48,7 @@ const Params = {
 
     // Debug logging
     debugging: false,
-    verboseLevel: 9,
+    verboseLevel: 8,
 
     // Status for sockets
     token_socket_connected: false,
@@ -61,7 +58,17 @@ const Params = {
     send_request_backlog_callbacks: [],
     logging_socket_callbacks: [],
     token_socket_callbacks: [],
+    
+    // locks for mutex
+    locks: {},
+    
+    token_request_queue: [],
 };
+
+Params.locks[Params.connection_socket.address().port] = false;
+Params.locks[Params.token_socket.address().port] = false;
+Params.locks[Params.logging_socket.address().port] = false;
+
 
 // Various connection types accepted by the server
 const ConnectType = {
@@ -84,13 +91,15 @@ const LoggingLevels = {
 const Flag = {
   NONE: 0,
   ALLOW_UNKNOWN_LOGGERS: 1,
-  AUTHORIZE_LOGGERS_WITH_NO_ACCESS_CODE: 4,
+  ALLOW_DEFAULT_ACCESS_CODE: 4,
   ALLOW_PLAIN_LOG_REQUEST: 8,
   ALLOW_BULK_LOG_REQUEST: 16,
-  COMPRESSION: 512
+  COMPRESSION: 256
 };
 
 const PACKET_DELIMITER = '\r\n\r\n';
+const DEFAULT_ACCESS_CODE = 'default';
+const PING_THRESHOLD = 60;
 
 // Utility static functions
 const Utils = {
@@ -130,7 +139,7 @@ const Utils = {
     now: function() {
         return parseInt((new Date()).getTime() / 1000, 10);
     },
-    
+
     getTimestamp: function() {
         return Utils.now();
     },
@@ -148,8 +157,9 @@ const Utils = {
         if (typeof compress === 'undefined') {
             compress = false;
         }
-        if (!nolock && Params.lock) {
+        if (!nolock && Params.locks[socket.address().port]) {
             Params.send_request_backlog_callbacks.push(function() {
+                Utils.debugLog('Sending request via callback');
                 Utils.sendRequest(request, socket, false, sendPlain, compress);
             });
             return;
@@ -163,22 +173,18 @@ const Utils = {
         }
         let encryptedRequest;
         if (!sendPlain) {
-            try {
-                let iv = new Buffer(crypto.randomBytes(16), 'hex');
-                let cipher = crypto.createCipheriv('aes-256-cbc', new Buffer(Params.connection.key, 'hex'), iv);
-                encryptedRequest = iv.toString('hex') + ':' + Params.connection.client_id + ':' + cipher.update(finalRequest, 'utf-8', 'base64') + cipher.final('base64') + PACKET_DELIMITER;
-            } catch (err) {
-                Utils.debugLog(err);
-            }
+            encryptedRequest = Utils.encrypt(finalRequest);
         } else {
             encryptedRequest = finalRequest + PACKET_DELIMITER;
         }
         Utils.vLog(9, 'Payload (Plain): ' + finalRequest);
-        Params.lock = true;
+        Utils.vLog(8, 'Locking ' + socket.address().port);
+        Params.locks[socket.address().port] = true;
         try {
             Utils.debugLog('Sending...');
             socket.write(encryptedRequest, 'utf-8', function() {
-                Params.lock = false;
+                Params.locks[socket.address().port] = false;
+                Utils.vLog(8, 'Unlocking ' + socket.address().port);
                 setTimeout(function() {
                     if (Params.send_request_backlog_callbacks.length > 0) {
                         const cb = Params.send_request_backlog_callbacks.splice(0, 1)[0];
@@ -187,25 +193,45 @@ const Utils = {
                 }, 10);
             });
         } catch (e) {
+            Utils.vLog(8, 'Unlocking ' + socket.address().port + ' [because of exception]');
+            Params.locks[socket.address().port] = false;
             Utils.debugLog('Error while writing to socket...');
             Utils.debugLog(e);
         }
     },
 
+    getCipherAlgorithm: function(keyHex) {
+      return `aes-${(keyHex.length / 2) * 8}-cbc`;
+    },
+
+    encrypt: function(request) {
+      let encryptedRequest;
+      try {
+          let iv = new Buffer(crypto.randomBytes(16), 'hex');
+          let cipher = crypto.createCipheriv(Utils.getCipherAlgorithm(Params.connection.key), new Buffer(Params.connection.key, 'hex'), iv);
+          return iv.toString('hex') + ':' + Params.connection.client_id + ':' + cipher.update(request, 'utf-8', 'base64') + cipher.final('base64') + PACKET_DELIMITER;
+      } catch (err) {
+          Utils.debugLog(err);
+      }
+      return '';
+    },
+
     // Decrypt response from the server using symmetric key
-    decrypt: function(response) {
+    decrypt: function(data) {
         if (Params.connection === null) {
             return null;
         }
         try {
-            const resp = response.toString().split(':');
+            const resp = data.split(':');
             const iv = resp[0];
             const clientId = resp.length === 3 ? resp[1] : '';
-            const base64Data = new Buffer(resp.length === 3 ? resp[2] : resp[1], 'base64');
-            let decipher = crypto.createDecipheriv('aes-256-cbc', new Buffer(Params.connection.key, 'hex'), new Buffer(iv, 'hex'));
+            const actualData = resp.length === 3 ? resp[2] : resp[1];
+            const binaryData = new Buffer(actualData, 'base64');
+            Utils.vLog(8, 'Reading ' + data.trim() + ' >>> parts: ' + iv + ' >>> ' + actualData.trim() + ' >>> ' + Params.connection.key);
+            let decipher = crypto.createDecipheriv(Utils.getCipherAlgorithm(Params.connection.key), new Buffer(Params.connection.key, 'hex'), new Buffer(iv, 'hex'));
             decipher.setAutoPadding(false);
 
-            let plain = decipher.update(base64Data, 'base64', 'utf-8');
+            let plain = decipher.update(binaryData, 'base64', 'utf-8');
             plain += decipher.final('utf-8');
             // Remove non-ascii characters from decrypted text ! Argggh!
             plain = plain.replace(/[^A-Za-z 0-9 \.,\?""!@#\$%\^&\*\(\)-_=\+;:<>\/\\\|\}\{\[\]`~]*/g, '');
@@ -216,12 +242,12 @@ const Utils = {
 
         return null;
     },
-    
+
     extractPublicKey: function(privateKey) {
         const key = new NodeRSA(privateKey.key);
         return key.exportKey('public');
     },
-    
+
     generateKeypair: function(keySize) {
         const key = new NodeRSA({b: keySize});
         key.setOptions({encryptionScheme: 'pkcs1'});
@@ -255,7 +281,7 @@ const Utils = {
 
 // Handle response from the server on connection requests
 Params.connection_socket.on('data', function(data) {
-    let decryptedData = Utils.decrypt(data);
+    let decryptedData = Utils.decrypt(data.toString());
     if (decryptedData === null) {
         decryptedData = Utils.decryptRSA(data, Params.rsa_key.privateKey);
     }
@@ -280,10 +306,12 @@ Params.connection_socket.on('data', function(data) {
         Utils.debugLog('Connecting to Residue Server...(step 3)');
         Params.connection = dataJson;
         Params.connected = true;
+        Utils.vLog(8, `Connection socket: ${Params.connection_socket.address().port}`);
         if (typeof Params.options.access_codes === 'object') {
             if (!Params.token_socket_connected) {
                 Params.token_socket.connect(Params.connection.token_port, Params.options.host, function() {
                     Params.token_socket_connected = true;
+                    Utils.vLog(8, `Token socket: ${Params.token_socket.address().port}`);
                     Utils.debugLog('Obtaining tokens...');
                     Params.options.access_codes.forEach(function(item) {
                         obtainToken(item.logger_id, item.code);
@@ -297,16 +325,18 @@ Params.connection_socket.on('data', function(data) {
             Params.logging_socket.connect(Params.connection.logging_port, Params.options.host, function() {
                 Utils.log('Connected to Residue!');
                 Params.logging_socket_connected = true;
+                Utils.vLog(8, `Logging socket: ${Params.logging_socket.address().port}`);
                 Params.connecting = false;
-
-                while (Params.logging_socket_callbacks.length > 0) {
+                const callbackCounts = Params.logging_socket_callbacks.length;
+                for (let idx = 0; idx < callbackCounts; ++idx) {
                     const cb = Params.logging_socket_callbacks.splice(0, 1)[0];
                     cb();
                 }
             });
         } else {
             Params.connecting = false;
-            while (Params.logging_socket_callbacks.length > 0) {
+            const callbackCounts = Params.logging_socket_callbacks.length;
+            for (let idx = 0; idx < callbackCounts; ++idx) {
                 const cb = Params.logging_socket_callbacks.splice(0, 1)[0];
                 cb();
             }
@@ -329,23 +359,38 @@ Params.connection_socket.on('error', function(error) {
 
 // Handle response for tokens, this stores tokens in to Params.tokens
 Params.token_socket.on('data', function(data) {
-    let decryptedData = Utils.decrypt(data);
+    let decryptedData = Utils.decrypt(data.toString());
     if (decryptedData === null) {
         Utils.log('Unable to read response: ' + data);
         return;
     }
     Utils.debugLog(decryptedData.toString());
-    const dataJson = JSON.parse(decryptedData.toString());
-    Utils.debugLog('Decoded json successfully');
-    if (dataJson.status === 0) {
-        dataJson.dateCreated = Utils.now();
-        Params.tokens[dataJson.loggerId] = dataJson;
-        while (Params.token_socket_callbacks.length > 0) {
-            const cb = Params.token_socket_callbacks.splice(0, 1)[0];
-            cb();
+    try {
+        const dataJson = JSON.parse(decryptedData.toString());
+        Utils.debugLog('Decoded json successfully');
+        if (dataJson.status === 0) {
+            dataJson.dateCreated = Utils.now();
+            Params.tokens[dataJson.loggerId] = dataJson;
+            const queuePos = Params.token_request_queue.indexOf(dataJson.loggerId);
+            if (queuePos !== -1) {
+                Params.token_request_queue.splice(queuePos, 1);
+            }
+            Utils.debugLog('New token: ');
+            Utils.debugLog(dataJson);
+            const callbacksCount = Params.token_socket_callbacks.length;
+            Utils.debugLog('Token callbacks: ' + callbacksCount);
+            for (let idx = 0; idx < callbacksCount; ++idx) {
+                Utils.debugLog('Token callback()');
+                const cb = Params.token_socket_callbacks.splice(0, 1)[0];
+                cb();
+                Utils.debugLog('Done Token callback()');
+            }
+        } else {
+            Utils.log('Error while obtaining token: ' + dataJson.error_text);
         }
-    } else {
-        Utils.log('Error while obtaining token: ' + dataJson.error_text);
+    } catch (e) {
+        Utils.log('Exception while obtaining token: ');
+        Utils.log(e);
     }
 });
 
@@ -365,15 +410,23 @@ Params.logging_socket.on('data', function(data) {
 
 // Obtain token for the logger that requires token
 obtainToken = function(loggerId, accessCode) {
-    if (!Params.connected || Params.connecting) {
-        Utils.log('Not connected to the server yet');
+    if (!Params.token_socket_connected) {
+        Utils.log('Not connected to the token server yet');
         return;
     }
-    if (Params.lock) {
+    /*
+    // after implementing token_request_queue lock check here shouldn't be needed!
+    if (Params.locks[Params.token_socket.address().port]) {
         Utils.debugLog('Already locked');
         return;
     }
-    if (typeof accessCode === 'undefined') {
+    */
+    if (Params.token_request_queue.indexOf(loggerId) !== -1) {
+        Utils.debugLog('Token already requested for [' + loggerId + ']');
+        return;
+    }
+    Utils.debugLog('obtainToken(' + loggerId + ', ' + accessCode + ')');
+    if (accessCode === null) {
         // Get from map (recursive)
         if (typeof Params.options.access_codes !== 'undefined') {
             let found = false;
@@ -386,24 +439,29 @@ obtainToken = function(loggerId, accessCode) {
                 }
             });
             if (!found) {
-                if (Utils.hasFlag(Flag.AUTHORIZE_LOGGERS_WITH_NO_ACCESS_CODE)) {
-                    Utils.debugLog('Trying to get token with no access code');
+                if (Utils.hasFlag(Flag.ALLOW_DEFAULT_ACCESS_CODE)) {
+                    Utils.debugLog('Trying to get token with default access code');
                     // try without access code
-                    obtainToken(loggerId, '');
+                    obtainToken(loggerId, DEFAULT_ACCESS_CODE);
                 } else {
                     Utils.log('ERROR: Access code for logger [' + loggerId + '] not provided. Loggers without access code are not allowed by the server.');
                     return;
                 }
             }
         } else {
-            if (Utils.hasFlag(Flag.AUTHORIZE_LOGGERS_WITH_NO_ACCESS_CODE)) {
-                Utils.debugLog('Trying to get token with no access code');
-                accessCode = '';
+            if (Utils.hasFlag(Flag.ALLOW_DEFAULT_ACCESS_CODE)) {
+                Utils.debugLog('Trying to get token with default access code');
+                accessCode = DEFAULT_ACCESS_CODE;
             } else {
                 Utils.log('ERROR: Loggers without access code are not allowed by the server');
                 return;
             }
         }
+    }
+    if (accessCode === null) {
+        // last hope!
+        Utils.debugLog('Forcing default access code');
+        accessCode = DEFAULT_ACCESS_CODE;
     }
     Utils.debugLog('Obtaining token for [' + loggerId + '] with access code [' + accessCode + ']');
     const request = {
@@ -411,13 +469,11 @@ obtainToken = function(loggerId, accessCode) {
         logger_id: loggerId,
         access_code: accessCode
     };
+    Params.token_request_queue.push(loggerId);
     Utils.sendRequest(request, Params.token_socket);
 }
 
-shouldSendPing = function(threshold/* = 60 */) {
-    if (typeof threshold === 'undefined') {
-        threshold = 60;
-    }
+shouldSendPing = function() {
     if (!Params.connected || Params.connecting) {
         // Can't send ping
         return false;
@@ -426,7 +482,7 @@ shouldSendPing = function(threshold/* = 60 */) {
         // Always alive!
         return false;
     }
-    return Params.connection.age - (Utils.now() - Params.connection.date_created) <= threshold;
+    return Params.connection.age - (Utils.now() - Params.connection.date_created) < PING_THRESHOLD;
 }
 
 sendPing = function() {
@@ -471,7 +527,7 @@ getCurrentTimeUTC = function() {
 }
 
 // Send log request to the server. No response is expected
-sendLogRequest = function(logMessage, level, loggerId, sourceFile, sourceLine, sourceFunc, verboseLevel) {
+sendLogRequest = function(logMessage, level, loggerId, sourceFile, sourceLine, sourceFunc, verboseLevel, callbackDepth) {
     if (Params.connecting) {
        Params.logging_socket_callbacks.push(function() {
             sendLogRequest(logMessage, level, loggerId, sourceFile, sourceLine, sourceFunc, verboseLevel);
@@ -484,12 +540,31 @@ sendLogRequest = function(logMessage, level, loggerId, sourceFile, sourceLine, s
         return;
     }
 
-    Utils.debugLog('Checking health...');
+    if (Params.token_request_queue.indexOf(loggerId) !== -1) {
+        Utils.debugLog('Waiting for token for logger [' + loggerId + '], requeueing...');
+        Params.token_socket_callbacks.push(function() {
+            Utils.debugLog('Sending log from requeued token callback... [' + loggerId + ']');
+            sendLogRequest(logMessage, level, loggerId, sourceFile, sourceLine, sourceFunc, verboseLevel, 1);
+        });
+        return;
+    }
+    /*
+    if (typeof callbackDepth === 'undefined') {
+        callbackDepth = 1;
+    }
+    
+    if (callbackDepth > 2) {
+        Utils.log('Ignoring log request from callback #' + callbackDepth);
+        return;
+    }*/
+
+    Utils.debugLog('Checking health...[' + loggerId + ']');
 
     if (!isClientValid()) {
         Utils.debugLog('Resetting connection...');
         Params.logging_socket_callbacks.push(function() {
-            sendLogRequest(logMessage, level, loggerId, sourceFile, sourceLine, sourceFunc, verboseLevel);
+            Utils.debugLog('Sending log from log callback... [' + loggerId + ']');
+            sendLogRequest(logMessage, level, loggerId, sourceFile, sourceLine, sourceFunc, verboseLevel/*, ++callbackDepth*/);
         });
         Params.connection_socket.destroy();
         disconnect();
@@ -497,25 +572,27 @@ sendLogRequest = function(logMessage, level, loggerId, sourceFile, sourceLine, s
         return;
     }
 
-    if (shouldSendPing(60)) {
+    if (shouldSendPing()) {
         Utils.debugLog('Pinging first...');
         Params.logging_socket_callbacks.push(function() {
-            sendLogRequest(logMessage, level, loggerId, sourceFile, sourceLine, sourceFunc, verboseLevel);
+            Utils.debugLog('Sending log from ping callback... [' + loggerId + ']');
+            sendLogRequest(logMessage, level, loggerId, sourceFile, sourceLine, sourceFunc, verboseLevel/*, ++callbackDepth*/);
         });
         sendPing();
         return;
     }
 
     if (!hasValidToken(loggerId)) {
-        Utils.debugLog('Obtaining token first...');
+        Utils.debugLog('Obtaining token first... [' + loggerId + ']');
         Params.token_socket_callbacks.push(function() {
-            sendLogRequest(logMessage, level, loggerId, sourceFile, sourceLine, sourceFunc, verboseLevel);
+            Utils.debugLog('Sending log from token callback... [' + loggerId + ']');
+            sendLogRequest(logMessage, level, loggerId, sourceFile, sourceLine, sourceFunc, verboseLevel/*, ++callbackDepth*/);
         });
-        obtainToken(loggerId);
+        obtainToken(loggerId, null /* means resolve in function */);
         return;
     }
 
-    Utils.debugLog('Sending log request...');
+    Utils.debugLog('Sending log request [' + loggerId  + ']...');
 
     let datetime = Params.options.utc_time ? getCurrentTimeUTC() : new Date().getTime();
     if (Params.options.time_offset) {
@@ -653,6 +730,7 @@ disconnect = function() {
             Utils.log('Disconnecting gracefully...');
             Params.token_socket.end();
             Params.logging_socket.end();
+            Params.tokens = [];
             Params.connected = false;
             Params.connection = null;
             Params.token_socket_connected = false;